ahora se envia el descuento del cliente.

+++ src/main/java/org/tallerjava/moduloClientes/dominio/Profesional.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 
 public class Profesional extends Cliente {
 
-    private Float porcentajeDescuento;
+    private double porcentajeDescuento;
     private EnumTipoProfesional tipo;
 
 }
